Add new optional parameter "fallback_method"
This will ensure that if we send an image to request a thumbnail
and it is smaller than the requested size it will not result in an ugly
zoom/streched image. Instead we will be able to choose a fallback method
like "resize" which results in a no-op in regards to image dimensions.