Use new API feature still_seek_percentage
This permits us to grab a still without guessing how many seconds to seek.
As a workaround for not having a way for users to pick a key frame / thumbnail,
we will avoid the blank still image problem by seeking to 50% of the video and
using that instead. It's not the best solution, but better than what we have.