Don't block competing requests for an item from the cache which is expired.
If we have a copy of the item in the cache, just serve it. If origin is offline
or gone but we have the media, why block clients trying to access it while we
wait on a response from origin, and why would we consider throwing it away?

Also ignore Expires header as extra precaution.